Configure HTML5 Mode
This commit will apply HTML 5 Mode routes to the application. This makes the application
much more SEO friendly and removes the bang operator from the address bar. To achieve this,
routes were also added to routes.rb so a user can reload the page and Angular will direct
them to the appropriate state.